:recycle:Refactor: 구현체로 구현받도록 구조 변경
